CDAP-14107 fix bug with RDD functions
Fixing a bug that was introduced during refactoring for the scala
spark sink where RDD functions were not being handled correctly.
In compute plugins, we were assuming the return value was a
DataFrame, and would end up calling .toJavaRDD() on a JavaRDD.
In sink plugins, we were assuming the return value was an RDD
instead of Unit, which would result in a null pointer exception.

Added correctly handling so that we only try to convert a DataFrame
to an RDD if it is actually a DataFrame, and we don't try to treat
a null as an RDD.